websocket: rename UserSockets to ClientManager
ClientManager manage all of active connection in server.
The ClientManager is part of the Server now, instead of separated instance.

+++ lib/websocket/websocket_test.go
@@ -81,7 +81,7 @@ func testHandleAuth(req *Handshake) (ctx context.Context, err error) {
 
 	ctx = context.WithValue(context.Background(), CtxKeyExternalJWT, extJWT)
 	ctx = context.WithValue(ctx, CtxKeyInternalJWT, _testInternalJWT)
-	ctx = context.WithValue(ctx, CtxKeyUID, _testUID)
+	ctx = context.WithValue(ctx, CtxKeyUID, uint64(_testUID))
 
 	return
 }
